Add tell-keeper script, Use Proper Style for ask-keeper
I realized when I linted ask-keeper for my job I needed to conform to
styles that were a little different from Python for Go.  I've rectified
my variable and function names to adjust for the style-required camel
case.

I also added the framework for a tell-keeper script to save a password
to Keeper.  I plan to add utilize both of these scripts in the
management of some secrets in build processes.

Keeper Secrets Manager has been pretty easy to write for and utilize via
the SDK auth method provided.